requirements.txt
PyJWT~=2.1.0
SQLAlchemy==1.4.32

flask~=2.1.2